Rename dss.storage.bundles to dss.storage.identifiers
