+++ dss/api/bundles/__init__.py
@@ -11,7 +11,7 @@ import requests
 from cloud_blobstore import BlobNotFoundError, BlobStore
 from flask import jsonify, request
 
-from dss.storage.bundles import TombstoneID, BundleFQID, FileFQID
+from dss.storage.identifiers import TombstoneID, BundleFQID, FileFQID
 from dss.util.blobstore import test_object_exists, ObjectTest
 from dss.util.bundles import get_bundle
 from dss.util.version import datetime_to_version_format
