+++ dss/util/bundles.py
@@ -5,7 +5,7 @@ from cloud_blobstore import BlobNotFoundError
 
 from dss import Config, DSSException, Replica
 from dss.storage.hcablobstore import BundleFileMetadata, BundleMetadata
-from dss.storage.bundles import DSS_BUNDLE_KEY_REGEX, DSS_BUNDLE_TOMBSTONE_REGEX, TombstoneID, BundleFQID
+from dss.storage.identifiers import DSS_BUNDLE_KEY_REGEX, DSS_BUNDLE_TOMBSTONE_REGEX, TombstoneID, BundleFQID
 from dss.util import UrlBuilder
 from dss.util.blobstore import test_object_exists
