+++ tests/test_indexer.py
@@ -28,7 +28,7 @@ from dss.config import Replica
 from dss.events.handlers.index import AWSIndexer, GCPIndexer, BundleDocument, Indexer
 from dss.storage.hcablobstore import BundleMetadata, BundleFileMetadata, FileMetadata
 from dss.logging import configure_test_logging
-from dss.storage.bundles import ObjectIdentifier, BundleFQID
+from dss.storage.identifiers import ObjectIdentifier, BundleFQID
 from dss.storage.validator import scrub_index_data
 from dss.util import create_blob_key, networking, UrlBuilder
 from dss.util.es import ElasticsearchClient
