+++ tests/test_object_identifier.py
@@ -8,7 +8,7 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noq
 sys.path.insert(0, pkg_root)  # noqa
 
 import dss
-from dss.storage.bundles import ObjectIdentifier, TombstoneID, FileFQID, BundleFQID, BUNDLE_PREFIX, FILE_PREFIX
+from dss.storage.identifiers import ObjectIdentifier, TombstoneID, FileFQID, BundleFQID, BUNDLE_PREFIX, FILE_PREFIX
 
 
 @testmode.standalone
